refactor(api): PR-5 commit 1 — api/ 骨架 + server.py 切入口 + lifespan 迁移
studio/api/ 新子包 8 文件接管 server.py 的 lifespan / middleware / 应用工厂
/ 启动入口 / 通用 helper，server.py 4657 行 → 顶部 110 行删，剩 130 routes
+ 4 工具 re-import。

  api/__init__.py       11 行  模块说明
  api/app.py            21 行  FastAPI(lifespan=...) + middleware
  api/lifespan.py      157 行  ensure_dirs/db.init_db + supervisor 启停 + SSE 计时器 + Proactor filter
  api/middleware.py     26 行  _SelectiveGZipMiddleware（/api/events + /samples/* 跳过）
  api/errors.py         63 行  _safe_join_or_400 / _validate_component_or_400 / data_export_path / _export_result
  api/responses.py      19 行  EMPTY_STATE 常量
  api/static.py         30 行  SPAStaticFiles（react-router 兜底 index.html）
  api/main.py           30 行  main() uvicorn 启动入口

**关键行为变更**：`ensure_dirs()` + `db.init_db()` 从 server.py 顶层
（import-time 副作用）移到 lifespan startup —— `from studio.server import
app` 不再触发文件系统初始化 / sqlite 表 init，便于测试 / 工具 import 而
不写盘。tests 在 fixture 里显式调 init_db() 的不受影响（多次 init_db()
幂等）。

server.py shim 策略：
- 顶部 `from .api.app import app`（共享同一 FastAPI 实例，130 route
  decorator 仍注册到本 app）
- 4 个 path helper + EMPTY_STATE 从 api.errors / api.responses re-import，
  保 server.py 内 35 处 callsite 零改动
- SPAStaticFiles 从 api.static re-import，mount 在 server.py 末尾不动
- main() 从 api.main re-export，`from studio.server import main` 兼容

行为零变更（除 lifespan 迁移）：所有 routes 仍在 server.py，签名 / 路径
/ 返回结构都没动；route_snapshot.py 显示 160 routes 不变；test_studio_server
44 routes 测试全过；inference daemon / supervisor / system_stats sampler
通过 lifespan 启停顺序与原 server.py 完全一致。

回归：PR-1 安全网 152/152 全绿（route snapshot + import smoke + invariants）
+ test_studio_server 44 + test_presets_io 20 + test_train_endpoints 6 +
test_version_config 8 + test_system_stats 9 等 = **267 全绿**。
